farm: route the reverse proxy by hub serial; WebSocket upgrade bridge
Implements the serial-routing first slice (farm-impl.md "Serial routing
— first slice"; decisions.md 2026-07-05). /hub/{serial}/{*path} now
resolves the segment as the hub's pubkey (unique partial index on
hubs.hub_pubkey) instead of the opaque id PK — clients already hold the
serial from farm-ready invite links. Distinct errors preserved:
404 hub_not_found / 503 hub_suspended / 503 hub_not_running; the
injected header is now x-hub-serial.

Upgrade requests can't ride the buffered reqwest path, so they branch
to a raw socket bridge: replay the request head to the hub, relay its
response, and on 101 run copy_bidirectional between the upgraded
client stream and the hub TCP stream. Verified end-to-end with a
tungstenite client against a stub WS echo hub.

Also fixes two latent process_port decode bugs (INTEGER column decoded
as i64 fails sqlx's runtime type check): the proxy silently 404'd
every registered hub — pre-existing and invisible because the proxy
route had zero test coverage — and spawn_all_from_db would have failed
the same way at farm startup. 5 new integration tests in
serial_routing_flow.rs.

UDP voice is deliberately not proxied; /info passes through unmodified
so clients will read the hub's voice endpoint from it once the hub
publishes voice_udp_addr (future slice, hub-side).

+++ crates/farm/src/hub_manager.rs
@@ -113,7 +113,9 @@ impl HubManager {
     /// Re-spawn all non-suspended, non-deleted hubs from the DB.
     /// Called once at farm startup.
     pub async fn spawn_all_from_db(&self, db: &PgPool) -> Result<()> {
-        let rows: Vec<(String, String, i64, Option<String>)> = sqlx::query_as(
+        // process_port is INTEGER (i32) — decoding as i64 fails sqlx's type
+        // check at runtime (same latent bug the proxy had; see proxy.rs).
+        let rows: Vec<(String, String, i32, Option<String>)> = sqlx::query_as(
             "SELECT id, db_path, process_port, owner_pubkey FROM hubs
              WHERE suspended_at IS NULL AND deleted_at IS NULL AND process_port IS NOT NULL",
         )
